+++ .eslintrc.json
@@ -32,6 +32,8 @@
     "@typescript-eslint/no-use-before-define": ["error"],
     "react-hooks/rules-of-hooks": "error",
     "react-hooks/exhaustive-deps": "warn",
+    "react/prop-types": "off",
+    "react/jsx-props-no-spreading": "off",
     "no-underscore-dangle": "off",
     "react/jsx-filename-extension": [1, { "extensions": [".tsx"] }],
     "import/prefer-default-export": "off",
