+++ src/pages/Dashboard/styles.ts
@@ -1,10 +1,5 @@
 import styled from 'styled-components/native';
 
-interface MovieItemProps {
-  screenWidth: number;
-  distanceToSelectedScroll: number;
-}
-
 export const Container = styled.SafeAreaView`
   flex: 1;
   background: #15151c;
@@ -32,19 +27,6 @@ export const MovieListHeaderText = styled.Text`
 
 export const MovieList = styled.View`
   padding-top: 32px;
-`;
-
-export const MovieItem = styled.View<MovieItemProps>`
-  width: ${(props) => props.screenWidth - 128}px;
-  transform: scale(
-    ${(props) =>
-      0.8 +
-      (props.distanceToSelectedScroll >= 1
-        ? 0
-        : (1 - props.distanceToSelectedScroll) * 0.2)}
-  );
-  margin: 0 8px;
-  align-items: center;
 `;
 
 export const MovieImage = styled.Image`
